Extract ColorPickerPopper to separate file
No functionality change.
